requirements.txt
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0
python-dotenv
flask-talisman
flask-limiter
flask-caching==2.1.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
